pyproject.toml
[project]
name = "trading-algo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fyers-apiv3>=3.1.7",
    "kiteconnect>=5.0.1",
    "matplotlib>=3.10.5",
    "mibian>=0.1.3",
    "numba>=0.61.2",
    "pandas>=2.3.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "ratelimit>=2.2.1",
    "requests>=2.31.0",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "stumpy>=1.13.0",
    "termcolor>=2.4.0",
]
